adding changes to the service page
